refactor(frontend): group provider connect files
